refactor: renamed folder menu to home
relates to #2
