+++ src/home/intro/birds/Birds.tsx
