+++ src/rootSaga.ts
@@ -1,6 +1,6 @@
 import { all } from 'redux-saga/effects';
 
-import { watchItemsRequest, watchPopularItemsRequest } from './menu/sagas';
+import { watchItemsRequest, watchPopularItemsRequest } from './home/sagas';
 
 export default function* rootSaga() {
   yield all([watchItemsRequest(), watchPopularItemsRequest()]);
