+++ src/styles/_components.scss
@@ -1,4 +1,4 @@
 @import '../common/header/Header';
 @import '../common/logo/Logo';
-@import '../menu/Menu';
+@import '../home//Home';
 @import '../page-not-found/PageNotFound';
